+++ samples/advanced/.mockserverrc.js
@@ -2,15 +2,16 @@
 // Install - npm install @r35007/mock-server
 // Run in Terminal - mock-server
 
+/** @type {import('@r35007/mock-server/dist/cli/index').CliOptions} */
 module.exports = {
   root: __dirname, // or process.cwd() // Root path of the server. All fetch paths in db data will be relative to this path
   db: "./db.json", // Path to database
   injectors: "./injectors.json", // Path to injectors
   middlewares: "./middlewares.js", // Path to middlewares
   rewriters: "./rewriters.json", // Path to rewriters
-  store: "./store.json",// Path to store
+  store: "./store.json", // Path to store
   snapshots: "snapshots", // Path to snapshots
-  static: "public", // Path to host a static files. Set to empty string to avoid hosting default public folder 
+  static: "public", // Path to host a static files. Set to empty string to avoid hosting default public folder
 
   port: 3000, // Set Port to 0 to pick a random available port.
   host: "localhost", // Set host to empty string to set your local ip address
@@ -36,9 +37,10 @@ module.exports = {
   ignoreFiles: [], // Paths to ignore file changes
   openInside: true, // Open Home page inside vscode
   showInfoMsg: true, // Show vscode notification on start and stop of server
-  statusBar: { // Status bar settings
+  statusBar: {
+    // Status bar settings
     position: "Right",
     priority: "0",
-    show: "true"
-  }
+    show: "true",
+  },
 };