+++ samples/advanced/server/detailed.js
@@ -1,30 +1,23 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 // Install - npm install @r35007/mock-server
 
-const path = require("path");
 const { MockServer } = require("@r35007/mock-server");
 // const MockServer = require("@r35007/mock-server").default; // For default import
 
-const config = {
-  root: path.resolve(__dirname, "../"), // All fetch paths will be relative to this path
-  port: 3000, // Set Port to 0 to pick a random available port. default: 3000
-  host: "localhost", // Set empty string to set your Local Ip Address 
-  quiet: false, // Set to true to suppress console logs
-  log: false // Set to true to see setter logs. If quiet is false the we cant see the setter logs.
-};
+const config = require("../.mockserverrc.js");
 const mockServer = MockServer.Create(config);
 
 const app = mockServer.app;
 
 // Sets global injectors, middlewares, store and rewriters
 mockServer.setData({
-  injectors: "./injectors.json",
-  middlewares: "./middlewares.js",
-  store: "./store.json",
+  injectors: config.injectors,
+  middlewares: config.middlewares,
+  store: config.store,
 }); // pass mockServer instance to use it in middleware.js method
 
 // Make sure to use this at first, before all the resources
-const rewriter = mockServer.rewriter("./rewriters.json");
+const rewriter = mockServer.rewriter(config.rewriters);
 app.use(rewriter);
 
 // Returns the default middlewares
@@ -45,9 +38,12 @@ app.use((req, res, next) => {
 app.get("/echo", (req, res) => res.jsonp(req.query));
 
 // Creates resources and returns the express router
-const resources = mockServer.resources("./db.json");
+const resources = mockServer.resources(config.db);
 
-resources.create("/todos", (req, res, next) => { next(); }) // /todos will be added to existing db
+resources
+  .create("/todos", (req, res, next) => {
+    next();
+  }) // /todos will be added to existing db
   .send({ userId: 1, id: 1, title: "Marvel", completed: false })
   .delay(1000) // in milliseconds
   .done(); // make sure to call done method at last to complete the route configuration